rename stuff into swervedrive
